Add initial website content and jobs
Add an initial website based on the arcana template from html5up,
which is CC-BY licensed.

It includes some initial content about Zuul, but is mostly intended
to bootstrap the process of collaboratively editing the site.  All
of the content in this repo should generally be CC-BY licensed as
well.

The banner image was created by me in my personal capacity, and it
is also licensed CC-BY.

This includes jobs to publish draft and final versions of the website,
including the production secret needed to publish to afs.

Change-Id: I994fea62061acb34037de1da27e90f2a09df4e74
Depends-On: https://review.openstack.org/538345