WIP Use gatsby for static site generation
The gatsby static site generator is a pure javascript/html based static
site generator. It allows content for the site to be written in markdown
and then compiled through a build step.

This also shifts to using fontawesome via the webpack build instead of
vendored, and it removes jquery which wasn't being used anyway.

A followup change will add netlify-cms as a through-the-web CMS for
content editing that will submit changes as patches to gerrit. The use
of gatsby is a step in enabling that, but the two are not tied to each
other, so simply updating what we have now to be in gatsby seemed like a
good first choice.

Change-Id: I35a46c7e72cd9b1a655569b95d94fae48ccf5a42