Add a promotional message banner and events list
During lead up to AnsibleFest in Austin, display a message letting
readers know that the Zuul team will be represented along with some
related talks/sessions, and include a link to additional information
about the event. Also add a section to the Community page listing
this and the upcoming OpenStack Summit in Berlin. Following
AnsibleFest, the banner can be updated with OpenStack Summit details
instead.

Change-Id: I7e58aaff69bd4425c0f3b32200907c036f8110f1