Add a "start" redirect
In support of print media (slides and brochures) use a short "start"
URL which currently redirects to the ZFS tutorial but can also be
repointed wherever we like if something better comes along. Update
the Get Started button on the index page to use this URL so we don't
need to change it in two places.

Change-Id: I0edd947ace425dd5bd009f9055b269dbbbec9d30